refactor(wallet): Share the wallet-data layer under UIContext, not UIContext
UIContext is the one that does not move, and this is the honest version of
sharing it.

The two copies model two different UIs. The extension coordinates a popup with
a full-page view -- RefreshMode, openBrowserWindow over chrome.tabs, an
openBrowser channel, pendingAuth/Credential/Alias arriving as props, plus its
own AuthContext. react-wallet has one window and its own deep-link pendings:
pendingChallenge, pendingSubTab, pendingHeldDID. Merging those would produce a
provider mostly made of questions about which host it is running in, which is
worse than two files.

Underneath them, though, the wallet-data layer was the same code. Of the
fifteen functions the two share by name, ten were byte-identical -- including
refreshAliases, at 105 lines the most intricate thing in either file -- and
three more differed only by a variable name. Those thirteen are now
useWalletData in the shared package, reading the keymaster and writing the
variables context, both of which are already shared.

refreshAll and resetCurrentID stay app-local on purpose: both reach into their
host (restore stored state, clear it, poke the browser view) and both are short
enough to compose from what the hook returns.

UIContext: 663 -> 371 lines here, 799 -> 509 there, with 272 and 270 lines of
duplicated logic deleted respectively.

So tier C ends 2/3 shared plus this: VariablesProvider and WalletProvider moved
whole, UIContext splits. The components that depend on these contexts are now
unblocked either way, since both halves keep the same hook names.

Refs #894
